🙈 ignore(.gitignore): ignore build files **/lib and minified file Makefile.js
Update the .gitignore file to exclude the build files in the "lib" directory from version control. This is achieved by adding the "**/lib" pattern, which will ignore all "lib" directories within any subdirectory of the project.

Additionally, the minified file "Makefile.js" will be ignored to prevent it from being included in version control. This ensures that the original Makefile remains the source of truth and avoids versioning unnecessary build artifacts.