📦 chore(tsconfig.json): add 'swc': true
In the tsconfig.json file, add the "swc": true configuration option. This setting enables the usage of @swc/core as the TypeScript compiler, allowing for faster and more efficient transpilation of TypeScript code. By enabling "swc" in the tsconfig.json, the project leverages the benefits of the @swc/core compiler for improved development performance.