🔨 script: add script to create Makefile.js
Create a script to generate the Makefile.js file in the project. This script will automate the process of creating the Makefile.js with predefined configurations for building the project. The Makefile.js will include rules for transpiling TypeScript to JavaScript using the @swc/core compiler, as well as other build-related tasks. This script will simplify the setup process and ensure consistency in the Makefile.js generation across different development environments.