🙈 ignore(.gitignore): ignore stryker-tmp folder
Add "stryker-tmp" to the .gitignore file to prevent it from being tracked in version control. Ignoring the "stryker-tmp" file ensures that any temporary files generated by the stryker mutator during mutation testing are not included in the repository. This helps maintain a clean and organized version control history and prevents unnecessary commits of temporary build artifacts.