🙈 ignore(.gitignore): add reports folder
Ignore the "reports" folder in the .gitignore file. This folder is automatically generated by Stryker for reporting the mutation status during mutation testing. By adding "reports" to .gitignore, the generated reports will be excluded from version control, ensuring that they are not committed to the repository. This helps keep the repository clean and prevents unnecessary commits of automatically generated files.