📦 chore: add eslint 8.46.0 style linter
This commit adds the necessary ESLint configuration file (eslintrc) to the project. The eslintrc file contains rules and settings for ESLint, defining how code should be linted and formatted. With this configuration in place, ESLint will be able to enforce coding standards and catch potential issues during development, promoting code consistency and quality.