✨ feat(bin/interkapto): add interkapto bin
Added the "bin/interkapto" script, which determines the execution mode (development or production) for the "interkapto" package.

Description:
The "bin/interkapto" script is responsible for determining the execution mode of the "interkapto" package based on the existence of the "src" directory. The script performs the following actions:

Checks if the "src" directory exists to determine if the script is being executed in development mode.
If in development mode:
Sets the "NODE_ENV" environment variable to 'development'.
Sets the path to the "tsconfig.json" file.
Registers the "ts-node" compiler with specific options, including "tsconfig-paths/register".
Requires the entry point file from the "src" directory.
If in production mode:
Sets the "NODE_ENV" environment variable to 'production'.
Requires the entry point file from the "lib" directory.
This script facilitates the execution of the "interkapto" package, adjusting the behavior based on the environment (development or production).